feat: Implement tag parsing (close #2)
find_all_unique_tags()
Returns an array of strings containing all unique tags in a note.

find_and_replace_all_tags()
Takes text and a Record of tags as input, returns text with the tags replaced with their associated value in the record or TAG_NOT_FOUND.

Implemented a MarkdownPostProcessor to replace tags with values in Reading View. If a tag is found, it is replaced and colored green, if not, it is colored red.